Moved the Audit Information button off of the MatchEditor and into its own MatchActionNode in the tree.
Audit Information is no longer displayed in a dialog but as an EditorPane on the right hand side. Since this class does not try to edit the displayed information, the methods in the EditorPane interface that warn against returning bogus values return bogus values. This should be ok as long as we don't start editing the information this 'editor' displays.