Changed the MatchMakerEngine interface so that all of the engine logging information is handled via a Log4J Logger instead of a pair of OutputStreams.
The core part of this change is completed and working, but the GUI portion
has not been completed yet.  We're about to arrange for some sort of Swing
component that displays Log4J messages from a certain logger.  Until then,
the output panel in the GUI will remain blank while the engine runs.