Fixed a bug where performing the findAll method on the a MatchPool multiple times (ex. when updating the display value data in the match validation screen) would duplicate all the PotentialMatchRecords. As a result, we would see a dashed-edge being painted over all edges, which would be most noticeable when trying to set edges to NO MATCH.
The duplication was because the when the findAll method is run more than once, it adds new PotentialMatchRecords, including copies of existing ones into a Set, and the old PotentialMatchRecord was using the default equals implementation (by reference), so it would accept them. As such, we had to re-implement the equals and hashcode methods.