-Added context menu to Merges Folder for adding new merge rule.  The new merge rule's editor is shown when the new merge rule is added. -Fixed a bug dealing with hibernate for saving merge rules by not making a duplicate. -Added documentation for related files and for r923. -Modified equality methods to better reflect a equality.
One outstanding bug still exists with hibernate when a new merge rule is added while there is an existing rule dealing with the same table.  Ideally, the merge rule editor's Table name list should remove any tables that already have existing merge rules.  Additionally, the mappings for merge criteria is currently a bag.  For the ordering of merge rules, it should be modified to a list.  As such ordering of merge rules is not currently reflected in the TreePath or the MergeTableRuleEditor.

Another abnormality with the merge rule editor is that when collision criteria actions are saved as Ignore, it would show up as UNKNOWN if the session was closed and reopened.  Not sure if it was originally intended.

Lastly, currently when new match rules and merge rules are added, the TreePath is not selected.  Ideally, it would be selected.