Changed the way translate words are persisted within their group:  Now we're using a Hibernate list mapping instead of a bag, because Hibernate will just use the translate word indices within the group word lists as the seq_no column value in the database.
This lets us get rid of all the machinery around the syncChildrenSeqNo stuff, and it also means we're not carrying the list position redundantly as a property of each translate word.

This change unfortunately has a down side: Now, if there are gaps in the sequence numbering in the database, Hibernate will insert corresponding null-valued list entries into the lists of translate words.  We have an old test that checks for this problem, and it's currently failing.  We plan to compensate by having the app present a good error message about the list gaps, so that at least users know what the problem is and how to fix it.  It might even be a reasonable idea to make a utility method for running through the list and removing the null items after reading a translate group from the database. (in the DAO)