Changed the way children of the Match are persisted.  Now match and merge rules are persisted and cascaded in the same way as each other, which is a good idea for consistent behaviour and easy code maintenance.
We modified the test case for moving rule sets from one parent match to another, becuase this has proven tricky under the new scheme and we were experimenting to find a working approach.  It's still not working, but Ben and Kevin are going to make it work by either saving a copy of the match rules instead of reassigning parent pointers, or by changing the child cascade styles to not automatically delete orphans.

Also in this commit:

-There was some lacking documentation and bad code placement in the MergeTableRuleEditor. We cleaned it up.

-Removed the seqNo properties from the merge rules, since they are already being handled as a list by Hibernate.