Changed the MungeStep Interface to allow the user to find the number of inputs exected (and a constant to return for unlimited). Added a method to find the type of input expected for each input. These were added so that connections could be checked when they are made.
Changed the MungeStepOutput equals method to rely on direct references so that in the user interface they can be used in a hash map to find there UI component.

Update some of the merge steps  to conform to the new interface.

Updated the test case stub to match the interface and moved it to the approprate package.