New step implementation that gets its data from a SQL query, and advances the query result set by one row on each call().
Currently, you have to give this step a SQLTable directly when you create it, but in the long run, it will have to look for the input table in its ancestor Match object.  I haven't implemented that yet because before we can do that, we have to get rid of the existing MatchRule class and substitute it for MungeStep in the MMO hierarchy.