Removed all references I could find to the old C engines.
Re-purposed the AbstractCEngine class to be just an AbstractEngine, since all it really does is implement Monitorable and Callable, and check a bunch of preconditions that still apply to our planned pure Java engines.

As of now, the MatchMaker is engineless!  We'll be committing the new match engine as soon as we get it written.