Changed the IOconnection so that it can paint itself, when passed in the graphics from the mungePen to make the code simpler and allow the possibility of the lines to be tabed through in the future.
In the process of switching to allow for the tabbing to switch between the components to improve usability. 

TODO: 
Right now the user needs to click twice sometimes for select a component.
Clean up the recentlyselected variable, right now it is used as a buffer for when the component has not had its focus set yet.