Changed the buildUI method in the MungeComponent so that it returns the panel as oppose to modifying it. This was done to make it easier for the user to build components.
Change MungeComponent to AbstractMungeComponent to better reflect what it is now that it has an abstract method.

Continued to improve the focus based selection of the components to allow for proper tabbing.

TODO:
Add keylisteners to the component because they are now intercepting the key events.

Fix one bug when a component is opened.

Clean up some parts of the code.

Add the ability to tab through lines.