Changed the IOConnection around so that it is now a JComponent and maintains its own bounds and handles its own events.
It now continuously sets it bounds and draws itself.

Removed the list of lines from the munge pen as this is now handled by the component superclass.

Each line is now a mouse listener for the munge pen so that they can properly handle the click events of the lines, before if a line was clicked within the bounds the top most line would steal the event even if there was one under it.

Changed the munge component so that the window will immediately redraw the scroll pane after a component is moved outside the frame.

Re-added the right click menu to the lines.

Munge Pen is now in a good condition.