Added MungeResultRuleSet to the set of munge components to visualize where the place where the user finishes with their data.
Fixed a bug where the bounds on the IOCs were changed and it did not remove parts of the line that are no longer on the bounds. 

Changed the input munge step to start with all of the inputs small to initially keep the component small so that it is easier to use.

Added Object and Data to the supported types in the GUI.