Moved the remove all listeners method in MungePen to be a static method so that it is no longer repeated in Abstract Munge Component and IOConnection.
Added fancy icons to the munge component.

Added pretty plus icons the the SQLInput munge component with a cool rollover effect.

Added the icons that will be needed to draw the plugs to their own packages.

Removed the set border command from the input and output munge components so that they will work with the new amazing look.

Changed the paint method for abstract munge component to draw the top part for the exciting new look, and added the new beautiful icons to the edge.

TODO:

Add the super new plugs to the munge components.