Added a drag and drop toolbar for the right hand side of the munge pen. This allows the user to add mungeComs in a logical way to the mungepen without memorizing seeming random choices in the buttons 1 - 0 and q - e.
The mungeStep list is locked in a folder in the right side because a scroll pane was not working well with the munge pen. And allowing the toolbar to move did interesting things to the size of the munge pen.

Added a getMungeStep method to the SwingContext so that a new munge step can be created with the information in the properties file.

Modified the StepDescription to extend Component because only components are handled by the transfer handler. 

Created a munge step library class that is responsible for doing the majority of the dragging operations (except the adding the component in the drop into the munge pen). Also it is resonpsible for creating the list of munge steps.

Gave the Input and Output Munge Components better default positions so they don't start on top of each other.

TODO:

Create a +/- type button to only show the munge step's icons.