Added a button to the munge step toolbar to allow it to shrink down to the size of the icons so that it takes up less space but is still functional.
Stole the chevron images for the button because they looked cool.

Added a default icon to the munge steps so they don't disipear when the tool bar shrinks if they don't have an icon. The color wheel was chosen for no perticular reason.

Got rid of the mouse listener on the munge pen to add steps because it is no longer nessary.