Moved the code that generates the output labels for the starting step to the munge component.
Modified it to allow any component to show its input or output names by calling setInputShowNames(true) and setOutputShowNames(true) respectivly.

Moved the hide/show all buttons there as well and they can now be used by any subclass.

Cleaned up the concat step so that it uses a simple flow layout instead of a form layout because it was overcomplicated and caused issues during testing.

Cleaned up to InputStep to simply use the new code in the abstract munge component.