Fixed an issue where the MatchPool was storing match results from different munge processes but for the same source records. The problem is that the match result table has a unique key constraint on the source record keys. This was particularly noticeable when running a match on multiple munge processes which would result in two source table records being matched in two or more munge processes.
To fix this, I modified PotentialMatchRecord's equals method to compare only on source record keys. This wouldn't have been an issue in the old engine anyway, since the old engine's results were from the database, so we would always get only one match per pair of source table records anyway.

Some additional code was also needed to over write a potential match in the match pool if adding a potential match with a higher match confidence. 

One consequence of this is that one of the Fudger tests will fail, since it is expecting both a fudge match and a real match to exist between two source records in the test match pool. I chose to ignore this for now since the Fudger will be removed after the new Merge engine is complete.