Reasonably large code bomb
Added a logger to be passed around through all the mungeSteps and process that will write to the panel at what ever level it feels like. Now it is possible to view all the mungeing as it happens in the engine.

Changed All the abstract munge step to accept a logger in the call method to be able to write out all data. 

Changed all the munge steps and tests to accept the logger

Added a field in the engine to accept which logging level to have the logger use.
Added a combobox in all the engines to allow the user to select said logging level.
Changed the engines to set the aformentioned level.