Added the ability to cancel an engine mid run because that is useful if the user made a mistake.
Now the Engines and the Processors periodicaly check the monitorable object they have to see if they have been canceled from the UI.

The match pool is now passed in an Aborter class, which is an inner class of the MatchProcess that it can check to see if the match process has been aborted. All this class does is check the monitor.

The EngineWorker now has an attached listener that is uses to notify the RunEngineAction when it is finished. And the RunEngineAction calls run on a passed in runnable to notify what ever created it. At also has a runnable for starting the engine.

The three panels now have abort buttons for obvious reasons.