Fixed a "too clever by half" problem where the feature intended to clear out the WHERE clause when you switch tables actually causes the WHERE clause to always start off blank, even if there is one in the Project instance.
(The problem was with the delayed initialization of the catalog, schema, and table combo boxes... they caused the event to trigger and wipe out the WHERE clause).

Solution: The WHERE clause no longer gets deleted when you switch tables.  This isn't as bad as it sounds now that you can't switch the input table after having saved the project.