Added the ability to choose a separate database for the pl schema and the table you are working on. This was done because only being able to work on a table in the same database as the repository is silly.
Note: If not datasource is specified it will go with the one the one the pl schema is in.

TODO: Allow the result table and source table to be in separate databases.

This is not an option now because there are a few messy joining statements that are not yet delt with.

What happened:

There is a place to store the spDataSource name of the database the user selected in the plschema, for the source result xref and table merge rules. right now they should always be the same, or merging wont work.

The SQL Object chooser was updated to hold all of datasources not just the current.  

The source talbe DataSource Chooser was added to the ProjectEditor, and the result tables DS chooser was rigged to be synced with the source one but not shown. That way the result combo boxes will be correct.

The FormVailidation handler was improved to have an option that it will not validate any components if they are disabled. This was added to stop the project editor from preventing you from saving if you have not selected an option that is not avaible for you DB of choice.

Updated the hybernate xml to update the spdatasource for the various fields.

Added an option in the cachced table to set the name of the data source that the table belongs to, so it can load a table from anywhere.

Updated the findPhysical table by name and table exists to optionally use the spdatasource name to find the right database.

Added a method in the session to get a database from a datasoruce. This was done so that all databases could be cached, if they were not cached and just created every time things like tables and columns would think they are not  equal when they are because they came from different database instances.

Added methods to the project to get and set the spdatasource name.

Created a methods in the project to create a connection to the source and target databases. This is done because getConnection was no longer good enough. 

Changed most of the engines/process ect to use the create connection methods as needed so they were talking to the right database.

Added a check to ensure that the given data source exists and throws an error if it does not.

Testing:

Changed the testing session context to immediatly add the three datasources to the list of datasources because it is now searched throught to find the right connection.

Added a test to ensure that the spDataSource name is persisted for the projects

Added two new test classes that test doing operations on separate databases.