Stopped the match maker from hordding connections. Now it will always hold 2 connections when idle, and maxes out at 4 (very briefly) when running the match engine (if the source table is in the PLSchema database).
This was caused by the temp DB's generated by getPhysicalTableByName reopening their connection to look at stuff.
This was solved by pre-populating part of the table when you do a find.

Changed the project save method to add the new result table back to the database as a child if needed to prevent there from being a reference to a temp DB when the table is retrieved.

Fixed the methods in the 3 different session classes (they should really be looked into).

Removed the MultiPlatformMungeProcessorTest because it never needed an data from the pl schema and was redundant.