Changed the startup process to automatically log in to a local HSQLDB repository instead of presenting a login prompt when the app starts.
Additionally, if the local repository database does not yet exist, it will be silently (and quickly!) created.

There are provisions for automatic repository update support, but this is not yet implemented (should be done before releasing a MatchMaker that uses a newer schema version)

Also pulled the "Create Repository" action out into its own action class (it used to be in the login dialog).  This action will be added to the Databases menu in the near future.