Finished implementing the auto-login preferences, including the GUI editor.  Now you can choose auto-login vs start with login dialog.  If you choose auto-login, you can even choose which repository to auto-login to.
The prefs action isn't hooked into the UI for non-mac yet (mac preferences appear under the special application menu).. I'll probably put it under the File menu like in the Architect.

Also found and removed an incorrect matchmaker app version constant in the session context.  Now the only place the version is tracked is in the MatchMakerVersion class.