Changed the Abstract munge step to be more functional. Now open, call, commit, rollback, and close are final and they call a corrosponeding do method that is defaulted to nothing. Now you don't need to worry about called super dot something.
Changed all of the munge steps to use this system.

Changed the input munge step input and cleanse result step to not require a project to be passed in. So every new munge process is added to the tree as it's created and the process is saved. This is done so that it will always have a project and so that discard changes will look like the tree, without this you could have a process in the tree that would not get persisted.

Uppercase munge step was not really deleted, it will be back in a second (eclipse SVN problem)