Now using the library's ConnectionComboBoxModel for the data source combo box.  Using the custom model means the combo box contents track the actual contents of the data source collection.
The upshot of all this is, if you add a new connection when the project editor is already visible (or anything else that uses the SQLObjectChooser), it will magically be in the combo box.  Before this change, you would have to leave the editor and come back to see the new connection.