Fixed a bug that prevents the matchmaker's changes from HSQL to be written to the disk. Now the shutdown=true is appended to the end of the jdbc connection so that HSQL will close itself and we will stop losing data.
This is added a a check when loading the default schema to make sure that shutdown=true.

Also closed the repository database in the close action of the match maker. This did not solve the problem with HSQL but is likely a good thing to have.

Added a regression test for it.