Fix for bug #1546.
The problem was that DuplicatePanel was creating the internal panel object as a separate field from the one inherited from NoEditEditorPane. Since the 'getPanel()' method returns the one from NoEditEditorPane, which would still be null, it returns a null panel which inevitably leads to a NPE. 

So instead of setting DuplicatePanel's own panel field, I use the 'setPanel()' method to set the inherited panel field.