Initial commit for a Maven Project Object Model setup for the Project Planner. Maven expects project modules to in their own subdirectories, so I took the liberty of creating these subdirectories, and src and regress folders in each. We can then gradually move the Project Planner classes into their respective folders.
project-planner-base: For shared classes and libraries between the client and server components.

project-planner-client: The client side Java WebStart application that the user will start up from their web browser

project-planner-server: The 'middleware' server that will perform the direct database operations and control access to the database from the clients. 

Note that the pom.xml files committed are not yet complete and will probably not work as of yet with Maven, but I will fill them out soon as I better understand how to get Maven to work with this project.