Created a new XML-based DAO for projects and all their metadata.  Currently, this is done in a monolothic manner, so you have to save/load a whole project at a time.
The purpose of this new DAO is twofold: First, it works for project import/export (sharing and backup); second, it can be used as a client-server communication layer. Then we can make clients that don't connect to the repository database at all.

This commit contains what I believe is a reasonably complete XML output routine. Input is not yet implemented (coming soon!).

There are no unit tests yet; the plan is to do reflection-based testing of modify-save-load-compare, like we do in the Power*Architect. The big advantage to this approach is that the test will automatically cover new properties and remind us to update the DAO.