Made updates to the project-planner-base/pom.xml so that the project can be built and executed using Maven. There's still a lot of things I need to figure out with Maven, since it expects our project to be configured in a certain way (this has to do with Maven's 'convention over configuration' philosophy).
I also moved most of the source over to project-planner-base for the time being, since we haven't figured out how to actually split the code at this point. (A disadvantage of us basing this on the matchmaker I think) When the time comes, we can move the source files into the project-planner-client and project-planner-server directories as needed.