Looks like the project-planner-base/pom.xml didn't get committed last time. Here's the comment from before
Made updates to the project-planner-base/pom.xml so that the project can be built and executed using Maven. There's still a lot of things I need to figure out with Maven, since it expects our project to be configured in a certain way (this has to do with Maven's 'convention over configuration' philosophy).