Looks like this fix got committed earlier without comment. I've added a comment to explain the fix.
The problem was when the client was running with Web Start, it would not be receiving a set-cookie header with every response as the WebsiteIOHandler was expecting, resulting in IllegalStateExceptions getting thrown on almost any operation requiring HTTP communication with the server. It looks like the client under Web Start would send a request with the 'Cookie' parameter set, which it didn't do under Eclipse.

This condition I added is just to prevent the IllegalStateException from being thrown needlessly. However, we may need to ensure that this condition doesn't prevent the IllegalStateException (because of set-cookie header not being found) from being thrown in situations where it should be.