Few changes to fix undo on labels.
For the parameter colour, it was calling setColour(String) but the actual method has Color in the signature.
Removed parameters for text area size since it wasn't needed.
Corrected MungePen in dealing with inserted munge steps on setting layers.
Removed reload() from executing on undo events as it was not needed and was causing an error in label's undo.