For LabelMungeComponent: * Looks like the color changer dialog was being kept in memory when generating thumbnails on the server, because it called pack() but did not dispose itself after. This resulted in the memory leak problem when generating thumbnails. I've prevented the setColorChanger method from being called until it's actually needed to prevent the dialog from calling pack() during the thumbnail generation. I also made the color changer static since I do not see any reason for each LabelMungeComponent to have their own color changer.
Additionally, I added a colorChanger.dispose() call in the remove method. In the future, it may be better for all MungeComponents to have a 'clean()' method which they can use to clean up after themselves (Like LabelMungeComponent disposing of its color changer), in order to help prevent future memory leaks such as this.

For ProjectPlannerServerUtils:
* Added an imageBuffer.flush() call. Not sure if it helped the memory leak problem, but it doesn't seem to hurt.