Debug Mode in the match engine should now rollback changes.
I've currently disabled the clear match pool option in debug mode because changes should be rolled back, but if the clearing of the match pool is rolled back but the engine thinks that it is cleared, it can cause unique key violations when it tries to insert 'new' matches when it calls pool.store() later on. But if the engine is made aware of the rollback, then it would be as if clear match pool wasn't selected in the first place, so I don't see the point in allowing it in debug mode for now.