Initial commit for the MatchMaker to automatically fix a project if its source table has been modified.
Added new checkPrecondition on the engines to verify source table structure. A new exception (SourceTableException) is thrown for changes in the source table. The user will then be prompted for the MatchMaker to try to fix the project.
Updated the MatchMakerEngine and its hierarchy's method signatures to handle the new SourceTableException.
Moved the code that creates a result table from the ProjectEditor to the MMSUtils so it could be used else where as well.

Factored out the checks on result table to be non static.
Fixed a bug in the ProjectEditor where the unique index combo box does not start with the project's index.
Changed the ProjectEditor so that when a user changes the unique index of a project, the source table merge rule updates to that change.