An initial commit of the new merge rule deriver. This is to fix an issue where the old merge rule deriver skipped tables which changed the imported column's name.
Instead of deriving merge rules by column names, it uses the actual foreign key constraint so that it doesn't matter if the column name is changed in the child table. Additionally, it also derives rules for child tables that import on alternate keys. Also, it will not mistakently create merge rules for tables that are not actually related to the source table but has a column with the same name as the source table's primary key column. (ex, if all tables used 'id' as a primary key name)

The option to use either or both merge rule derivers is available, since the new deriver will not work if the data model does not have foreign key constraints defined (either by mistake or poor design, or using a DB that doesn't support it, like MySQL with the MyISAM storage engine)

There are some remaining issues that will need to be fixed:
1) The validator for the column merge rules editor used to check that the number of imported columns equalled the size of the primary key. I've commented that out for now because of two problems. a) It doesn't account for alternate key imports and b) it does not account for the fact that the primary key can be imported multiple times.

2) The new merge rule deriver's performance on our Oracle XE test instance is pretty slow (then again, so are a lot of things running off of that Oracle instance). A Netbeans profile indicates that vast majority most of the time is being spent in database metadata decorator calls.

3) There is an issue if a child table inherits from more than one parent table. The particular scenario I've run into is when a parent table exports a key to two child tables, and then one of those child tables (Let's call it Table A) exports its primary key to the other child table (Call it Table B).

What the merge engine cannot currently handle is if the Table B is merged first, and then Table A gets merged, but in doing so, performs operations that require furthing merging in Table B again (ex. Deletes a record in Table A that is still referenced by a record from Table B). The problem here is that Table B is only aware of one of its parents, and it seems the merge engine is dependent on this to determine each table merge rule's children.

I currently cannot think of an easy way to solve this. So far, I can only think of having TableMergeRules support multiple parent TableMergeRules. However, I am not sure if the MatchMakerObject business model, or the current Hibernate storage configuration can support this. Something I'll have to look into.