Brought in the XML read and write code from the Project Planner branch. This makes it possible to both export and import one or more projects to/from an XML file.
In order to support the XML reading, we had to make several formerly private and protected methods public. We have made notes in the javadoc for those methods that they are not supposed to be used by normal client code.

Although we presently write out merge rules, the code for reading them back in is not yet complete. We will be working on that today.