Further improvements to the XML import/export based on a new test case.
This test is similar to the ones we use for the Hibernate persistence layer in that it uses reflection to find and modify the values of properties. This gives us some insurance that we will remember to update the file format when we add new properties to the objects we save and load.

The test is not yet complete, but as-is, it passes. We will now expand the scope of the test and the export/import code to try and make both complete.