The XML DAO test should now cover everything except merge rules, and it's passing!
While working on this, I found a problem with the MungeStep interface: it didn't have a way of asking for the number of inputs a step has. Now it does.

Another change I made was to pull out the FakeSQLDatabase class from the SQLInputStepTest so that the new XML DAO test could use it too. I started to realize we were reimplementing this again inside another test, and did the refactoring up front instead of next year. :)