Fix for bug #1649.
The problem was that even if the user defined a unique index, when opening the project editor, it only populates the unique index combobox with only the existing unique indexes on the source table, leaving out the user-defined index.

But, there is one problem here. If the selected unique index existed in the table and then it got removed, then when loading the project editor, it would get added in as a 'user-defined index'. It's not necessarily a bad thing, but the user should probably be made aware that the unique index doesn't officially exist in the table anymore. We will need to deal with this when expanding the scope of the 'check source table for changes' functionality.