AbstractMungeStep now has a check that returns true if it has at least one Input that has a MungeStepOutput connected to it, and false if it does not. This check is used to check if the result step in a munge process has any inputs coming into it. This is to prevent useless match and cleanse engine runs, and also prevent the match engine from running in such a circumstance, which results in every record matching with every other record. Not a very useful result, and the match visualizer slows down to a crawl as a result.
Also added a new test for this method, and modified the MungeProcessorTest to include a result step now. Also modified the MockJDBCCleanseTest to ignore the exception that gets thrown by this check, since it appears to be making sure the cleanse engine doesn't do anything to the source data if its result step is not connected.