Updated the old Table Explorer to the more generic and useful query tool in the library.
The DBTree used is part of the architect db lib, which was added to the classpath. This tree does not currently support drag and drop as the similar tree in Wabit does, as the SQLObjectSelection transferable class needed to be able to do this is dependent on the Architect. It would be a good idea to move that class to the library.