Cleaned up the abstract munge component. Removed redundant panels and now use the panel's setVisible method to hide and show parts of the components instead of adding and removing them to expand the components.
This has the side effect of fixing a bug where going to the munge 
pen would stack all the input and output connectors on top of each
other at the far left of each component.
This also has the side effect of fixing the bug where the Google
Address Lookup will start showing its content but once the expand
button is pressed the content will not be possible to get at again.