* Added the ability to start an 'Address Correction Project'. The project type itself is still more or less a clone of a deduping project so it doesn't really do anything yet. It's 'result' table is structured differently however. It's also worth noting that the result table structure is likely to have more columns added as development of the address correction project progresses.
* I removed the cross reference project options from the GUI, since it is not implemented, I don't see any point in teasing people with buttons for it that only say 'sorry, it's not implemented'. 

* Modified the test cases to ignore the addressCorrectionEngine for persistence. 

* Added a stub implementation for an AddressCorrectionEngine. 

* Updated the architect_lib since it includes some new DDL Generator classes. Since the new pl.ini format (which comes with the latest sql power library which gets bundled with the matchmaker) references these new classes, the library should be updated to prevent fun NoClassDefFoundErrors from happening.

I should note that I'm not all that fond of how Project has been implemented. Using a flag to change behaviour seems to be the wrong approach, especially coming from the OOP way of doing things. I've noticed it's been like this since the early days of MatchMaker. I don't want to refactor it just yet though, due to some deadline constraints, but I would like to do a refactoring of it at some point in the near future.