Added placeholders in the post_create.sql repository script so that the catalog and schema of the repository can be specified to the SQL statements. This is particularly useful if you created your repository in the non-default schema.
For example, in PostgreSQL, if you didn't create your repository in the 'public' schema, then originally, those insert statements would be called on the 'public' schema by default, but with the tables created in another schema, those statements would fail. This should prevent that problem.

There is another related problem though. After repository creation, attempting to log in will also fail because the code that looks up the user in the pl_user table also doesn't use a fully-qualified table name, resulting in it looking in the 'public' schema by default. If the repository wasn't created in the 'public' schema, then the matchmaker won't find it and login will fail.