I rolled back the existing utility class for refreshing a project, because it can only work if the database is modified while the MatchMaker is still running. The new refresh feature will support refresh no matter what state the MatchMaker was in when the table structure changed.
In terms of new refresh progress, the SQLInputStep now automatically refreshes itself based on the current state of the project's SQLTable. This refresh happens in the step's doOpen() routine. There are a number of tests that demonstrate this refresh procedure working properly.

Still to do are: The cleanse result step, the project (source table unique identifier and result table), and the merge rules.