Some major progression with the underlying engine for AddressCorrection
One major change is the concept of passing in an EngineMode flag to the MungeSteps when opening them, which can cause their behaviour to change depending on an engine's mode. The main use for this is so that the AddressCorrectionStep will set its outputs based on the contents of the result table if the AddressCorrectionEngine is committing corrected results back into the source table. I've also added this flag to other steps if they find it useful in the future. 

The implementation is not entirely ideal. I wanted to use enumerations to act as the flag type, but I also didn't want to bunch all enumerations for all steps in one enumeration, and also allow for a generic type that all enumerations can extend. Unfortunately, emumerations cannot extend another because any enumeration already extends java.lang.Enumeration, and since Java does not allow multiple inheritance, Enumerations can only implement interfaces. To allow a generic type to pass in as a flag to steps, I had an EngineMode interface that's actually empty, but which new enumerations can implement and be declared in the steps that actually use them. I think this is preferable to bunching up all enumerations for all steps in one, but feel free to comment on that.

The Engine also pays attention to the recently added pool filter and auto validation settings, although it has not been thoroughly tested yet.