The Address Correction Munge Step should return some better results in the 'invalid or different format mode', in that it doesn't store results in which an address has only one suggestion, and that suggestion and the original source address are appear and are formatted exactly the same.
Also, the Address Correction Step when running in 'write addresses back into the source table' mode, will now check the new 'isValid' flag in the AddressResult. 

Also, made a modification to the AddressCorrectionMungeProccessor and munge step that was preventing it from doing SERP auto-correction or committing addresses.

Finally, the AddressCorrectionMungeStep's checkPreconditions method forgot to add the ValidateResult into the list it was returning.