This commit fixes a problem where the delivery installation type was being appended to addresses when there was no delivery installation name. This caused addresses to be appended with things like "PO".
The new address correction, also in this commit, will check if the 
address type parsed matches the address types of the postal code given
in the database. If the address type parsed from address line 1 is 
different from the address types given from the database for the
input postal code the validator will try to find a more correct postal
code which covers the same address type as that which was parsed in
address line 1.